fixes plotting and minor things, nb

+++ quantus/metrics/faithfulness/road.py
@@ -204,6 +204,6 @@ class ROAD(PerturbationMetric):
 
         # Calculate accuracy for every number of most important pixels removed.
         self.last_results = {
-            percentage: np.mean(np.array(self.last_results)[:, p_ix])
+            percentage: np.mean(np.array(self.last_results).astype(int)[:, p_ix])
             for p_ix, percentage in enumerate(self.percentages)
         }
